Switch to SQLite for the indices. This is much faster and cleaner.
There is still a lot to refactor and simplify, and we could move a lot more to the sqlite (which we could then distribute more easily than the pickles).

Lets not put the library in there for now, but the DOI, structures and taxa names can go there. And the parenting relations can as well.